fix: decode playback_rate as f64 to match DECIMAL(3,2) column
sqlx can't decode MySQL DECIMAL directly into f32 without the
rust_decimal/bigdecimal feature. Cast to DOUBLE in SQL (+ 0.0) and
decode into f64 instead, matching how playback_rate is already used
downstream (cast to f64 for clock_rate).

+++ src/models/score.rs
@@ -152,7 +152,7 @@ pub struct RippleScore {
     pub max_combo: i32,
     pub full_combo: bool,
     pub mods: i32,
-    pub playback_rate: f32,
+    pub playback_rate: f64,
 
     #[serde(rename = "300_count")]
     #[sqlx(rename = "300_count")]
